Refactoring out dependency on CDI test jar
