+++ modules/flowable-camel/src/test/java/org/flowable/camel/cdi/named/CdiCamelContextFactory.java
@@ -1,3 +1,15 @@
+/* Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.flowable.camel.cdi.named;
 
 import javax.enterprise.context.ApplicationScoped;
@@ -7,6 +19,10 @@ import javax.inject.Named;
 import org.apache.camel.CamelContext;
 import org.apache.camel.impl.DefaultCamelContext;
 
+/**
+ * 
+ * @author Zach Visagie
+ */
 public class CdiCamelContextFactory {
 
     @Produces
