+++ modules/flowable-camel/src/test/java/org/flowable/camel/cdi/named/NamedCamelCdiFlowableTestCase.java
@@ -1,24 +1,36 @@
+/* Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.flowable.camel.cdi.named;
 
-import org.flowable.cdi.test.BaseCdiFlowableTestCase;
+import org.flowable.camel.cdi.BaseCamelCdiFlowableTestCase;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
- * This test base class is for supporting named contexts. The default camel context in a CDI environment is not named. A factory class is provided to name and construct the context.
+ * This test base class is for supporting named contexts. The default camel
+ * context in a CDI environment is not named. A factory class is provided to
+ * name and construct the context.
  * 
  * @author Zach Visagie
  *
  */
-public abstract class NamedCamelCdiFlowableTestCase extends BaseCdiFlowableTestCase {
-    protected static final Logger LOGGER = LoggerFactory.getLogger(NamedCamelCdiFlowableTestCase.class);
-    
+public abstract class NamedCamelCdiFlowableTestCase extends BaseCamelCdiFlowableTestCase {
+
     @Deployment
     public static JavaArchive createDeployment() {
-        return ShrinkWrap.create(JavaArchive.class).addPackages(true, "org.flowable.cdi").addPackages(true, "org.flowable.camel.cdi.named").addAsManifestResource("META-INF/beans.xml", "beans.xml");
+        return ShrinkWrap.create(JavaArchive.class).addPackages(true, "org.flowable.cdi").addPackages(true, "org.flowable.camel.cdi.named")
+                        .addAsManifestResource("META-INF/beans.xml", "beans.xml");
     }
 
 }
