+++ modules/flowable-camel/src/test/java/org/flowable/camel/cdi/std/CdiCamelVariableBodyMapTest.java
@@ -1,3 +1,15 @@
+/* Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.flowable.camel.cdi.std;
 
 import static org.junit.Assert.assertEquals;
@@ -13,6 +25,8 @@ import org.apache.camel.LoggingLevel;
 import org.apache.camel.Route;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
+import org.flowable.camel.CamelVariableBodyMapTest;
+import org.flowable.camel.impl.CdiCamelBehaviorBodyAsMapImpl;
 import org.flowable.engine.runtime.ProcessInstance;
 import org.flowable.engine.test.Deployment;
 import org.flowable.task.api.Task;
@@ -20,6 +34,11 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
+/**
+ * Adapted from {@link CamelVariableBodyMapTest} to test {@link CdiCamelBehaviorBodyAsMapImpl}.
+ * 
+ * @author Zach Visagie
+ */
 public class CdiCamelVariableBodyMapTest extends StdCamelCdiFlowableTestCase {
 
     protected MockEndpoint service1;
@@ -29,7 +48,7 @@ public class CdiCamelVariableBodyMapTest extends StdCamelCdiFlowableTestCase {
 
     @Before
     public void setUp() throws Exception {
-    	super.setUp();
+        super.setUp();
         camelContext.addRoutes(new RouteBuilder() {
 
             @Override
@@ -52,7 +71,7 @@ public class CdiCamelVariableBodyMapTest extends StdCamelCdiFlowableTestCase {
     }
 
     @Test
-    @Deployment(resources = {"process/HelloCamelCdiBodyMap.bpmn20.xml"})
+    @Deployment(resources = { "process/HelloCamelCdiBodyMap.bpmn20.xml" })
     public void testCamelBody() throws Exception {
         Map<String, Object> varMap = new HashMap<>();
         varMap.put("camelBody", "hello world");
