+++ modules/flowable-camel/src/test/java/org/flowable/camel/cdi/std/StdCamelCdiFlowableTestCase.java
@@ -1,11 +1,21 @@
+/* Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.flowable.camel.cdi.std;
 
-import org.flowable.cdi.test.BaseCdiFlowableTestCase;
+import org.flowable.camel.cdi.BaseCamelCdiFlowableTestCase;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * This test base class is for supporting the default camel context which is not a named bean in a CDI environment.
@@ -13,8 +23,7 @@ import org.slf4j.LoggerFactory;
  * @author Zach Visagie
  *
  */
-public abstract class StdCamelCdiFlowableTestCase extends BaseCdiFlowableTestCase {
-    protected static final Logger LOGGER = LoggerFactory.getLogger(StdCamelCdiFlowableTestCase.class);
+public abstract class StdCamelCdiFlowableTestCase extends BaseCamelCdiFlowableTestCase {
 
     @Deployment
     public static JavaArchive createDeployment() {
