+++ modules/flowable-camel/src/test/resources/META-INF/services/org.flowable.cdi.spi.ProcessEngineLookup
@@ -1,0 +1,1 @@
+org.flowable.camel.cdi.CamelCdiTestProcessEngineLookup
