+++ modules/flowable-cdi/pom.xml
@@ -14,21 +14,6 @@
 		<version>6.3.0-SNAPSHOT</version>
 	</parent>
 
-	<build>
-       <plugins>
-           <plugin>
-               <groupId>org.apache.maven.plugins</groupId>
-               <artifactId>maven-jar-plugin</artifactId>
-               <executions>
-                   <execution>
-                       <goals>
-                           <goal>test-jar</goal>
-                       </goals>
-                   </execution>
-               </executions>
-           </plugin>
-        </plugins>
-    </build>
 
 	<profiles>
 		<profile>
